pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "bop-modeling-data-airflow"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "apache-airflow>=2.10.4",
    "apache-airflow-providers-snowflake>=6.0.0",
    "dbt-core>=1.9.1",
    "dbt-snowflake>=1.9.0",
]

[tool.setuptools.packages.find]
where = ["bop_airflow"]

[dependency-groups]
dev = [
    "pytest>=8.3.4",
    "ruff>=0.9.1",
]
